Move from covert.stack_hists to utils.sum_hists
