+++ src/heputils/__init__.py
@@ -1,6 +1,7 @@
 from .version import __version__
 from . import plot
 from . import convert
+from . import utils
 
 # Satisfy pyflakes
-__all__ = ["__version__", "plot", "convert"]
+__all__ = ["__version__", "plot", "convert", "utils"]
